# api/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Route Guide Gateway
    description: Interface exported by the server.
    version: 0.0.1
servers:
    - url: http://localhost:8081
paths:
    /v1/features:
        get:
            tags:
                - RouteGuide
            description: |-
                A server-to-client streaming RPC.

                 Obtains the Features available within the given Rectangle.  Results are
                 streamed rather than returned at once (e.g. in a response message with a
                 repeated field), as the rectangle may cover a large area and contain a
                 huge number of features.
            operationId: RouteGuide_ListFeatures
            parameters:
                - name: lo.latitude
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lo.longitude
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: hi.latitude
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: hi.longitude
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Feature'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/features/{latitude}/{longitude}:
        get:
            tags:
                - RouteGuide
            description: |-
                A simple RPC.

                 Obtains the feature at a given position.

                 A feature with an empty name is returned if there's no feature at the given
                 position.
            operationId: RouteGuide_GetFeature
            parameters:
                - name: latitude
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
                - name: longitude
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Feature'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        Feature:
            type: object
            properties:
                name:
                    type: string
                    description: The name of the feature.
                location:
                    $ref: '#/components/schemas/Point'
            description: A feature names something at a given point. If a feature could not be named, the name is empty.
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        Point:
            type: object
            properties:
                latitude:
                    type: integer
                    format: int32
                longitude:
                    type: integer
                    format: int32
            description: Points are represented as latitude-longitude pairs in the E7 representation (degrees multiplied by 10**7 and rounded to the nearest integer). Latitudes should be in the range +/- 90 degrees and longitude should be in the range +/- 180 degrees (inclusive).
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: RouteGuide